refactor(domain): gather domain
