--- src/components/Dendrogram/State.ts
@@ -1,4 +1,0 @@
-import * as TreeLink from "../TreeLink";
-import * as TreeNode from "../TreeNode";
-
-export type State = TreeLink.State & TreeNode.State;
