--- src/components/TreeLink/State.ts
@@ -1,6 +1,0 @@
-import * as Types from "@this/types";
-
-export interface State {
-  scale: Types.Adjustment;
-  links: Types.Link[];
-}
