--- src/components/TreeNode/State.ts
@@ -1,8 +1,0 @@
-import * as Types from "@this/types";
-
-export interface State {
-  scale: Types.Adjustment;
-  nodes: Types.Node[];
-  radius: number;
-  offset: number;
-}
