+++ src/components/TreeNode/index.ts
@@ -1,4 +1,3 @@
 export { Container } from "./Container";
 export { generateStore, Store } from "./Store";
-export * from "./State";
 export * from "./TreeNode";
