+++ src/domain/App/State.ts
@@ -5,4 +5,9 @@ export interface State {
   currentKey: string;
   pkg: Types.Package;
   scripts: string[];
+  scale: Types.Adjustment;
+  links: Types.Link[];
+  nodes: Types.Node[];
+  radius: number;
+  offset: number;
 }
