+++ src/domain/App/__tests__/calc.test.ts
